worked on Dao class

+++ sprint-1/14-review/budget-tracker/src/main/java/com/niantic/models/Transaction.java
@@ -60,9 +60,8 @@ public class Transaction {
         return amount;
     }
 
-    public void setAmount(double amount) {
-        this.amount = BigDecimal.valueOf(amount);
-    }
+    public void setAmount(BigDecimal amount) {this.amount = amount;}
+
 
     public LocalDate getTransactionDate() {
         return transactionDate;
